refactored components and root index file.  everything is nice and modular now
