+++ react-redux-goals-todos/src/components/App.js
@@ -1,11 +1,33 @@
-import React from 'react';
+import React from 'react'
+import { connect, Provider } from 'react-redux'
+import ConnectedTodos from './Todos'
+import ConnectedGoals from './Goals'
+import Loading from './Loading'
+import { handleReceiveData } from '../actions/shared-actions'
 
-function App() {
-  return (
-    <div className="App">
-      Hello World
-    </div>
-  );
+class App extends React.Component {
+  componentDidMount () {
+    const { dispatch } = this.props
+    dispatch(handleReceiveData())
+  }
+
+  render () {
+    const { loading } = this.props
+
+    if (loading) {
+      return <Loading />
+    }
+
+    return (
+      <div>
+        <ConnectedTodos />
+        <ConnectedGoals />
+      </div>
+    )
+  }
 }
 
-export default App;
+
+export default connect((state) => ({
+  loading: state.loading
+}))(App)
