+++ react-redux-goals-todos/src/components/List.js
@@ -1,0 +1,20 @@
+import React from 'react'
+
+export default function List ( props ) {
+  const items = props.items
+  return (
+    <ul>
+      {items.map((item) => (
+        <li key={item.id}>
+          <span
+            onClick={() => props.toggle && props.toggle(item)}
+            style={{textDecoration: item.complete ? 'line-through' : 'none'}}
+          >
+              {item.name}
+          </span>
+          <button onClick={() => props.remove(item)}>X</button>
+        </li>
+      ))}
+    </ul>
+  )
+}
