+++ react-redux-goals-todos/src/components/Loading.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Loading () {
+  return (
+    <h2>Loading...</h2>
+  )
+}
