+++ react-redux-goals-todos/src/index.js
@@ -1,6 +1,17 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
-import App from './components/App';
+import { Provider } from 'react-redux'
+import { createStore } from 'redux'
+import reducer from './reducers'
+import middleware from './middleware'
+import ConnectedApp from './components/App';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+const store = createStore(reducer, middleware)
+
+ReactDOM.render(
+  <Provider store={store}>
+    <ConnectedApp />
+  </Provider>,
+  document.getElementById('root')
+)
+
